Modified underlying client_class.py METHOD and internal functions to give the end-user better control over calling their REST APIs in case they don't follow the standard CRUD model.  This allows for calls like base_url+'/some_string_or_different_format/or_different_path_entirely'.
Added a feature to allow striping the trailing '/' from the API url if desired.

Added a feature to allow passing custom HEADERS with the API calls.

Modified, base functions to also allow passing a JSON body to an api call
instead of just parameters, as some APIs take strictly JSON.